Script changes for _BDSIM_
Installing _Geant4_ and _BDSIM_ involved a lot of configuration. There is a script called `checkversions.sh`, adapted from an older version used at UCL, that check the versions of all the prerequisites for these programs. This is now updated to work for both _Geant4_ and _BDSIM_ on my current system.  Also added a script and an alias for loading and running _BDSIM_.