Created `.bash_functions` for common functions, such as _Git_ checks
Created function `git-check-all` to loop 
through my _Git_ repositories and check the
current status.

Moved `git-check` from an alias to a function,
as it was getting unwieldy already and I 
wanted to add more spacing, and it is easier
to have `git-check` and `git-check-all` in 
the same place.

Includes simple utility function `pause`.