Wrote initialisation lists and Get and Set methods for bunch names
Initialisation lists are compulsory for constant members (I've now set `_bunchCount` to be constant) and advised for all members.

The Set methods for bunch names include a default method with no arguments that is used by the default constructors, and can be called by the user.

The main Set method adjusts the list of names to match the number of bunches by adding default names or removing extra names.