Replace vector of bunch counts with fixed size array
Vectors don't work for loading data into the tree directly, as they are 
not stored in memory with the rest of the elements of the structure.

Variable-length arrays don't work because structures can only have 
fixed-length arrays.

Created a `_MAX_BUNCH_COUNT` parameter for the array size, and currently 
set this to 99.