Theme switcher scripts for iTerm2
Two separate scripts, one to run at startup, the other to run as a daemon.

I couldn't get a single script to do both, because of the security system of iTerm2 that only allows each script to call the run functions once.

It might be nice to put the common code into a separate file, rather than duplicating it.