Change logic for `tmate_start` command
Previous logic was incorrect: it was meant to only restart the connection if it was not already running, but the code actually did the opposite.
The problem was probably masked by the fact that the test fails if the socket is not specified.

New logic keeps the `main` session alive if possible, but restarts the `backup` session without checking.
This should hopefully mean that the main session is always accessible and doesn't get interrupted, and the backup session is regularly restarted to make sure there is a connection available in case of problems with the existing sessions.
Hopefully this is the best of both worlds.