Fix `get_summary()` method for readme pages
If there were no blank lines after the summary (as is normally the case for readme pages) then the previous code only returned the first line of the summary.

Modified code to return the whole contents if there is no blank line after the summary.

If there are bullet points after the summary, these are already handled and removed.