Adjust notebook folder comparisons
Previously would only check files in the list of standard pages, but would be better to check all files in the generator folder.
This is helpful for testing the rebuild method for notebooks and logbooks.

This affects the `get_pages()` method tests, as these only return standard pages, hence the need to remove non-standard pages from the generator folder before testing the result.