Add methods to write changes to disk
Page files write to the file listed as its path, or create a new file from the filename and parent path.
Notebooks iterate over contents and call the save method.

Pages check whether they have been modified, and if there are no changes the files on disk are not overwritten.