Adjust navigation lines for logbook pages
The previous setup was a bit confused, with "up" links included with the "previous" and  "next" links.

Now the upper line includes a breadcrumb trail all the way to the current month in the logbook, showing the full context from top to bottom; and the lower line is confined to the current level, showing (previous, this, next) at the month or day level.